LayoutTests/http/wpt/service-workers/fetchEvent.https.html

<!DOCTYPE html>
<title>Service Worker Fetch Event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/WebKit/resources/gc.js"></script>
<script>
// FIXME: Should be run on a service worker.
test(() => {
    var event = internals.createBeingDispatchedFetchEvent();
    event.respondWith(undefined);
    assert_throws('InvalidStateError', () => event.respondWith(undefined));
}, "FetchEvent respondWith should throw if called twice");

test(() => {
    var event = new FetchEvent('FetchEvent', { request : new Request('test') });
    event.request.value = 1;
    gc();
    assert_equals(event.request.value, 1);
}, "FetchEvent request is SameObject");

promise_test(async t => {
    if (!window.internals)
          return Promise.reject("test require internals");
    var event = internals.createBeingDispatchedFetchEvent();
    var promise = internals.waitForFetchEventToFinish(event);
    event.respondWith(undefined);
    return promise_rejects(t, new TypeError, promise);
}, "FetchEvent should be in error if responding with undefined");

promise_test(async t => {
    if (!window.internals)
        return Promise.reject("test require internals");
    var event = internals.createBeingDispatchedFetchEvent();
    var promise = internals.waitForFetchEventToFinish(event);
    event.respondWith(new Request(''));
    return promise_rejects(t, new TypeError, promise);
}, "FetchEvent should be in error if not responding with a Response");

promise_test(async t => {
    if (!window.internals)
        return Promise.reject("test require internals");
    var event = internals.createBeingDispatchedFetchEvent();
    var promise = internals.waitForFetchEventToFinish(event);
    event.respondWith(new Promise((resolve, reject) => {
        resolve(new Request(''));
    }));
    return promise_rejects(t, new TypeError, promise);
}, "FetchEvent should be in error if responding with a Promise that does not resolve to a Response");

promise_test(async t => {
    if (!window.internals)
        return Promise.reject("test require internals");
    var event = internals.createBeingDispatchedFetchEvent();
    var promise = internals.waitForFetchEventToFinish(event);
    event.respondWith(new Promise((resolve, reject) => {
        reject('not good');
    }));
    return promise_rejects(t, new TypeError, promise);
}, "FetchEvent should be in error if responding with a Promise that rejects");

promise_test(async t => {
    if (!window.internals)
         return Promise.reject("test require internals");
    var event = internals.createBeingDispatchedFetchEvent();
    var response = new Response;
    event.respondWith(response);
    assert_true(response === await internals.waitForFetchEventToFinish(event));
}, "FetchEvent should resolve when responding with a Response");

// Duplicate test from WPT. To be removed once WPT service worker tests are active.
test(function() {
    var req = new Request('http://localhost:8800/',
                          {method: 'POST',
                           headers: [['Content-Type', 'Text/Html']]});
    assert_equals(
      new ExtendableEvent('ExtendableEvent').type,
      'ExtendableEvent', 'Type of ExtendableEvent should be ExtendableEvent');
    assert_throws(new TypeError, function() {
        new FetchEvent('FetchEvent');
    }, 'FetchEvent constructor with one argument throws');
    assert_throws(new TypeError, function() {
        new FetchEvent('FetchEvent', {});
    }, 'FetchEvent constructor with empty init dict throws');
    assert_throws(new TypeError, function() {
        new FetchEvent('FetchEvent', {request: null});
    }, 'FetchEvent constructor with null request member throws');
    assert_equals(
      new FetchEvent('FetchEvent', {request: req}).type,
      'FetchEvent', 'Type of FetchEvent should be FetchEvent');
    assert_equals(
      new FetchEvent('FetchEvent', {request: req}).cancelable,
      false, 'Default FetchEvent.cancelable should be false');
    assert_equals(
      new FetchEvent('FetchEvent', {request: req}).bubbles,
      false, 'Default FetchEvent.bubbles should be false');
    assert_equals(
      new FetchEvent('FetchEvent', {request: req, cancelable: false}).cancelable,
      false, 'FetchEvent.cancelable should be false');
    assert_equals(
      new FetchEvent('FetchEvent', {request: req, clientId : 'test-client-id'}).clientId, 'test-client-id',
      'FetchEvent.clientId with option {clientId : "test-client-id"} should be "test-client-id"');
    assert_equals(
      new FetchEvent('FetchEvent', {request : req, isReload : true}).request.url,
      'http://localhost:8800/',
      'FetchEvent.request.url should return the value it was initialized to');
  }, 'Event constructors');
</script>
